==> JupyterScripts/00a_color_settings.ipynb <==
import seaborn as sns

c1 = sns.color_palette("Set1")
c1

c2 = sns.color_palette("Set2")
c2

c3 = sns.color_palette("pastel")
c3

phenotypes = {
    'U'  :c1[8],
    'S'   :c1[1],
    'A_r' :c1[7] ,
    'B_r' :c1[4],
    'A&B' :c1[3],
    'AB_r':c1[2],  
    'Other':c1[6], 
}

antibiotics = {
    "none":c2[0],
    "A":c2[1],
    "B":c2[2],
    "AB":c2[3],
}

strategies = {
    "No treatment":c3[-1],
    "Mono A":c3[0],
    "Mono B":c3[1],
    "Combo":c3[2], 
    "Mixing":c3[3],
    "Cycling":c3[4]
}

from evaluation_scripts.base import dump_toml
dump_toml("phenotype_colors.toml", phenotypes)
dump_toml("antibiotic_colors.toml", antibiotics)
dump_toml("strategy_colors.toml", strategies)
