==> JupyterScripts/01a_Collect_instructions.ipynb <==
# TS-Plasmids Evaluation

%load_ext autoreload
%autoreload 2

from evaluation_scripts.strainplate import CreateStrainplate_layout_384
from evaluation_scripts.base import get_pathes, FileNames, get_t_end, multiply_setup_plates
from evaluation_scripts.instructions import Import_Raw_Instructions, get_turnover_id, infection_to_well_id, get_transfer_id, transfer_from_to_id
from evaluation_scripts.well_ident import Label_Replicates, build_dictionaries, makeID
import pandas as pd
import json
from pandas.io.json import json_normalize 

## Load basic information

jupyter_path, basepath, path, analysis_path, os = get_pathes("20220412_output")

additional_infos = os.path.join(path, "Notes", "info.json")
infos = {"strainplate_layout":"new", "setup":"new"}
if os.path.exists(additional_infos):
    with open(additional_infos) as json_data:
        i = json.load(json_data)
    infos.update(i)
print(infos)

## Load all names of ... files in folder ... in the experiment folder ... and drop name of lock files (appear when opened)
instruction_file_names = FileNames(".csv", path, "platefiles")

## Create Dictionaries with basic infos
hue_infos, strategies, plate_rows, agar_strain_dict = build_dictionaries()

if infos["setup"] == "new":
    #First copy t0_P1_i to t0_P2-6_i_setup_copy. This is the easiest way to get the setupplate data into every strategy
    multiply_setup_plates(path, os)

## Create basic strainplate + exceptions

if infos["strainplate_layout"] == "new":
    t_end = get_t_end(instruction_file_names)
    strainplate = CreateStrainplate_layout_384(range(0,t_end+1), plate_rows)
    strainplate = Label_Replicates(strainplate, plate_rows)
    try:
        os.mkdir(analysis_path+os.sep+"obj")
    except:
        pass
else:
    from old_eval.ProperEvalFunctions import(CreateStrainplates_old_layout_384, CheckStrainPlates)
    t_end = get_t_end(instruction_file_names)
    strainplate = CreateStrainplates_old_layout_384(range(0,t_end+1), plate_rows)
    strainplate = CheckStrainPlates(path, strainplate)
    strainplate = Label_Replicates(strainplate, plate_rows)
    
    
obj_path = os.path.join(analysis_path, "obj")

if os.path.exists(os.path.join(analysis_path, "obj")) == False:
    os.mkdir(obj_path)
    
strainplate.to_pickle(obj_path+os.sep+"Strainplate.pkl")

## Write r-well infections/transfer/turnover to specific well ID
- Infection to
    - Notation not ideal ideal
        - The Instruction for t1, RWell 5 is to infect RWell 6
        - This means that RWell 6 is infected by Rwell 5 at t0 not t1...
    - Thats why the id instruction is written in the t0 well... 

## Just loads instructions into dataframe and sets correct index
print("Import_instructions")
Data = Import_Raw_Instructions(path,instruction_file_names)
# The transfer_to_well is actually wrong. It should actually be the transfered to r_well column... Otherwise they collide with turnover
Data = Data.rename(columns={"infection_to_well":"infection_to_rwell_with_time_error" })

## Adds which replicate is in which well
print("Label Replicates")
Data = Label_Replicates(Data, plate_rows)

## Get Turnover well IDs
print("Get turnover well id")
Data["turnover_id"] = Data.apply(lambda x:  get_turnover_id(strainplate, x) , axis=1)
Data = Data.sort_values(by=['transfer_n', "plate", 'rwell']).drop(columns=["transfer_to_well"])

Data[Data.transfer_n > 0]

## Correct infections_for_time
tmp = Data.loc[Data.infection_to_rwell_with_time_error.isnull()==False, ].reset_index().copy()
idx = tmp.apply(lambda x: makeID(x.transfer_n-1, x.plate, x.row, x.col), axis=1)
Data.loc[idx.values, "infection_to_well"] = tmp["infection_to_rwell_with_time_error"].values


## Calculate the Well IDs based on the r_wells, replicates, time and plates for transfer and infection
# Thats why it will be dropped
print("Get infection well id (to well)")
#Data = r_well_to_id_infection(Data)
Data["infection_to_well_id"] = Data.apply(lambda x: infection_to_well_id(x, Data), axis = 1)

print("Get transfer well id")
# Calculate where the well transfers to..
Data["transfer_from_well_id"] = Data.apply(lambda x: get_transfer_id(x), axis=1)
Data = Data.drop(columns=["infection_to_rwell_with_time_error"])

Data.apply(lambda x: transfer_from_to_id(x, Data), axis=1).head(50)

# save the data
Data.to_pickle(analysis_path+os.sep+"obj"+os.sep+"01a_data_instructions.pkl")
Data.head()
