==> JupyterScripts/01b_prepare_Data.ipynb <==
# TS-Plasmids Evaluation

%load_ext autoreload
%autoreload 2
import pandas as pd
import numpy as np
import pickle

from evaluation_scripts.base import get_pathes, FileNames, load_json
from evaluation_scripts.well_ident import build_dictionaries
from evaluation_scripts.phenotyping import (phenotyping,
                                            get_infection_strain, 
                                            get_transfer_strain, 
                                            get_turnover_strain,
                                            get_x
                                            )
from evaluation_scripts.data_preperation_1b import (exclude_unusable_plates,
                                                    added_wells,
                                                    added_strains, 
                                                    add_time_stamps)

experiments = ["20220412", "20220127", "20210417"]


for exp in experiments:
    print("\n", exp)
    
    ### LOAD
    jupyter_path, base_path, path, analysis_path, os = get_pathes(exp+"_output")
    #od_file_names = FileNames(".xml", path, "OD"+os.sep+"xml")
    hue_infos, strategies, plate_rows, agar_strain_dict = build_dictionaries()
    Data = pd.read_pickle(analysis_path+os.sep+"obj"+os.sep+"01a_data_instructions.pkl")
    strainplate = pd.read_pickle(analysis_path+os.sep+"obj"+os.sep+"Strainplate.pkl")
    exp_pars = load_json(os.path.join(path, "exp_pars.json"))
    control_wells = exp_pars["format"][0]["rwells_bl"]
    
    
    ### DO
    ## Add the results of the Agarplates for each well
    Data = exclude_unusable_plates(Data, path)
    Data = phenotyping(Data, path)
    
    ## Rename WT to S
    for c in ["phenotype", "turnover_strain"]:
            Data[c] = Data[c].replace({'WT': "S", 'wt': "S"})

    ## Track turnover phenotype and contamination
    Data["contaminated"] = False
    print("Track turnover phenotype and contamination")
    Data.apply(lambda x:  get_turnover_strain(x, Data, strainplate) , axis=1) 

    ## Track infection phenotype and contamination
    print("Track infection phenotype and contamination")
    Data["infected_by_wells"] = None
    Data["infected_by_strains"] = None
    Data.apply(lambda x:  get_infection_strain(x, Data) , axis=1)  

    ## Track transfer phenotype and contamination
    print("Track transfer phenotype and contamination")
    Data["received_transfer_strain"] = None
    Data.apply(lambda x:  get_transfer_strain(x, Data) , axis=1) 
    
    ## Add time stamps
    Data = add_time_stamps(Data, path)
    Data["dt"] = Data["t_replication_start"] - Data["t_turnover_start"]
    
    ## Collect added wells
    print("Collect added wells")
    Data["added_strains"] = Data.apply(lambda y: added_strains(y), axis = 1)
    Data["added_wells"] = Data.apply(lambda y: added_wells(y), axis = 1)
    
    ## Get x
    Data["x"] = Data.apply(lambda x: get_x(x.added_strains), axis = 1)

    
    ## Exclude contaminted wells
    Data.loc[(Data.contaminated == True),"exclude"] = True
    Data.loc[(Data.contaminated == True),"comment"] = "contamination"
    
    ## Exclude control wells
    Data.loc[Data.rwell.isin(control_wells), "exclude"] = True
    Data.loc[Data.rwell.isin(control_wells), "comment"] = "Control Well"
        
    ## Sort columns
    Data = Data.reindex(sorted(Data.columns), axis=1)
    Data[(Data.infected_by_wells != None) & (Data.infected_by_wells.isnull() == False)].head()
    
    ## Rename Combo to Combination
    Data.strategy.replace("Combo", "Combination", inplace=True)

    ### Save Data    
    Data.to_pickle(analysis_path+os.sep+"obj"+os.sep+"01b_Data.pkl")
Data.head()

Data.strategy.unique()
