==> JupyterScripts/02a_contaminations.ipynb <==
## Load
%load_ext autoreload
%autoreload 2
from evaluation_scripts.experiment_class import Experiment
from evaluation_scripts.transition_probabilities_class import Transition_probabilities
import pandas as pd
from itertools import product

experiments = ["20220412", "20220127", "20210417"]

dfs = []
for i, exp in enumerate(experiments):
    print("\n ", exp)
    
    experiment = Experiment(exp, exclude = False)
    df = experiment.data
    dfs.append(df)
df = pd.concat(dfs)
df = df[df.barcode.isnull() == False]

df["infected_well"] = df.phenotype != "U"
df["control"] = df.comment == "Control Well"
df["infected_control"] = df.control & df.infected_well

df.infected_control.sum()

df.control.sum()
