==> JupyterScripts/03c_transition_probabilities_av.ipynb <==
## Transition probabilities
- T describes: $\hat{x} = T x$
- we average over all experiments

## Load
%load_ext autoreload
%autoreload 2
from evaluation_scripts.experiment_class import Experiment
from evaluation_scripts.transition_probabilities_class import Transition_probabilities
from evaluation_scripts.base import save_for_latex
import pandas as pd
import os


experiments = ["20220412", "20220127", "20210417"]
experiment = Experiment(experiments)


## Do
- Get the x phenotypes.
    - S + U = S, A + A = A, A+B = A&B etc.
    - Equations can be found in the SI of the paper
- The loaded data correspond to $\hat{x}(t)$
- This data will be transformed to $x(t+1)$
- Than we can find a T that describes $\hat{x}(t+1) = Tx(t)$, since $\hat{x}(t+1)$ is known
- fishies will remain fishy exept they are mixed with ABr

print("create transition rates for t!=1")
t = Transition_probabilities(experiment, write = True)

print("create transition rates for t==1")
t1 = Transition_probabilities(experiment, t1 = True, write = True)


t.path

for antibiotic in t.antibiotics:
    path = os.path.join(t.pathes["base"], "tables", "M_"+antibiotic+".csv")
    M = t.transition_matrices["M_t_"+antibiotic]
    save_for_latex(M, [], path, round = 2)

for antibiotic in t1.antibiotics:
    path = os.path.join(t.pathes["base"], "tables", "M1_"+antibiotic+".csv")
    M = t1.transition_matrices["M_t1_"+antibiotic]
    save_for_latex(M, [], path, round = 2)

t.transition_matrices["M_AB"]

t.get_M(write = True)

from evaluation_scripts.base import load_pickle_obj
M_path = os.path.join(experiment.pathes["base"], "summary", "obj")
M_ab = load_pickle_obj(os.path.join(M_path, "M_AB.pkl")) 
M_ab

os.path.join(M_path, "M_AB.pkl")

t.transition_matrices["M_t_B"]
