==> JupyterScripts/03d_Loss_of_single_resistance_2.ipynb <==
## Special Skript for the 20220412 Experiment

import pandas as pd
import seaborn as sns
from matplotlib import pyplot as plt
import matplotlib.ticker as plticker
import pickle
import numpy as np
from evaluation_scripts.base import load_experiment, save_for_latex
import os

strains, exp_pars, dictionaries, pathes, df1, t_end = load_experiment("20220412")
_, _, _, _, df2, _ = load_experiment("20220127")
_, _, _, _, df3, _ = load_experiment("20210417")
df = pd.concat([df1, df2, df3])

df_b = df[df.x == "B_r"].copy()



def summerize(df, strain):
    cols = ["treatment_with", "plasmid_loss"]
    df_s = df[df.x == strain].copy()
    df_s["plasmid_loss"] = df_s.phenotype == "S"
    summary = df_s[cols].groupby(cols[:-1]).sum()
    summary["t"] =  df_s[cols].groupby(cols[:-1]).count()
    summary["f"] = round(summary.plasmid_loss/summary.t, 2)
    return summary
sum_a = summerize(df, "A_r")
sum_b = summerize(df, "B_r")
sum_b

plasmid_loss = pd.concat([sum_a.loc[:,["f"]], sum_b.loc[:,["f"]]], axis = 1)
plasmid_loss.columns = ["Ar", "Br"]
plasmid_loss.index.name = "treatment"
plasmid_loss = plasmid_loss.T
plasmid_loss

save_for_latex(plasmid_loss, [], os.path.join(pathes["general_obj"], "03d2_plasmid_loss.csv") )
