==> JupyterScripts/03e_clearance_probs.ipynb <==
import pandas as pd
import numpy as np
from evaluation_scripts.base import load_experiment, save_for_latex
import os

from evaluation_scripts.experiment_class import Experiment
experiment = Experiment(["20210417", "20220127", "20220412"])
data = experiment.data

## LOAD
strains, _, dictionaries, pathes, df1, _ = load_experiment("20210417")
_, _, _, _, df2, _ = load_experiment("20220127")
_, _, _, _, df3, _ = load_experiment("20220412")
data = pd.concat([df1, df2, df3])
data = data[data.exclude == False]
data.phenotype = data.phenotype.replace({"UI" : "U"})
path = pathes["general_obj"]

def get_clearance(data):
    df = data[ (data.x.isin(["S", "A_r", "B_r"])) & (data.treatment_with != "none") & (data.turnover_strain.isnull() == False)].copy()
    df["cleared"] = df.phenotype == "U"
    cols = ["x", "treatment_with", "cleared"]
    summary = df[cols].groupby(cols[:-1]).sum()
    summary["total"] = df[cols].groupby(cols[:-1]).count()
    summary["f"] = summary.cleared/summary.total
    summary = summary.reset_index()
    clearance = pd.pivot_table(summary, columns = "x", index = "treatment_with", values = "f")
    order = ["A", "B", "AB"]
    clearance = clearance.loc[order,]
    return clearance

clearance = get_clearance(data)

## SAVE
save_for_latex(clearance, [], os.path.join(pathes["tables"], "clearance_mean.csv"))

clearance
