==> JupyterScripts/03h_superconjugation_advanced_phenos.ipynb <==
## Special Skript for the 20220412 Emergence Experiment

%load_ext autoreload
%autoreload 2
import os
from evaluation_scripts.experiment_class import Experiment
import pandas as pd

date = "20220412"
experiment = Experiment(date, src="02g_advanced_phenotyping")
data = experiment.data

data[data.x == "A&B"].added_advanced_phenotypes

def sort_array(v):
    ## This function: 1) sorts the strains in a given order and 2) throws out multiple listings of same strain
    order = {
        0:"U^?",
        1:"S^h",
        2:"S^l",
        3:"S^?",
        4:"A_r^h",
        5:"A_r^l",
        6:"A_r^?",
        7:"B_r^h",
        8:"B_r^l",
        9:"B_r^?", 
        10:"A&B^?",
        11:"AB_r^?",
        12:"Fishy^?",
        13:"?",
        14:"Other^?"
    }
    order_inv = {v: k for k, v in order.items()}

    keys = []
    for s in v:
        keys.append(order_inv[s])
    keys.sort()
    res = []
    for k in keys:
        if order[k] not in res:
            res.append(order[k])
    return res
data["added_advanced_phenotypes"] = data.apply(lambda x: sort_array(x["added_advanced_phenotypes"]), axis = 1)

data[data.x == "A&B"].added_advanced_phenotypes

## Keep Only  Suitable Wells
- keep only wells where all advanced phenotypes are known
- keep only wells where somekind of A_r is mixed with somekind of B_r
- keep only wells that dont receive A&B, ABr or Other

data[data.x == "A&B"]

def check_suitability(row):
    v = row["added_advanced_phenotypes"]
    if len(v) == 2:
        b = True
        for vi in v:
            if ("?" in vi):
                b = False
        print(v)
        if (("A_r" in v[0] )& ("B_r" in v[1])) == False:
            b = False
        return b
    else:
        return False

def check_suitability(row):
    added_phenos = row["added_advanced_phenotypes"]
    if len(added_phenos) > 1:
        suitable = True
        a_there = False
        b_there = False
        
        for pheno in added_phenos:
            ## We can ignore Formes of S and U
            if ("S" in pheno) or ("U" in pheno):
                pass
            else:
                ## Some Phenotypes are not okay.
                if ("?" in pheno) or ("Other" in pheno) or ("AB" in pheno) or ("A&B" in pheno):
                    suitable = False

                ## Some A has to be there
                if (("A_r^h" in pheno) or ("A_r^l" in pheno)):
                    a_there = True
                    
                ## Some B has to be there
                if (("B_r^h" in pheno) or ("B_r^l" in pheno)):
                    b_there = True
        return suitable and a_there and b_there
    else:
        return False
interesting_combos = data[data.apply(lambda x: check_suitability(x), axis = 1)].copy()
interesting_combos

- save the incoming strains as s1 and s2 
- add a bool column if superconjugation happend

def reduce_pheno(added_phenos, pheno):
    result = [element for element in added_phenos if pheno in element]
    if pheno+"_r^h" in added_phenos:
        return pheno+"_r^h" 
    else:
        return pheno+"_r^l" 

def reduce_added_strains(added_phenos):
    return [reduce_pheno(added_phenos, "A"), reduce_pheno(added_phenos, "B")]
    
interesting_combos["superinfection"] = interesting_combos.added_advanced_phenotypes.apply(lambda x: reduce_added_strains(x))

def get_strain_i(row, i):
    return row.superinfection[i]
interesting_combos["conjugation"] = interesting_combos.phenotype == "AB_r"
interesting_combos["s1"] = interesting_combos.apply(lambda x: get_strain_i(x, 0), axis = 1)
interesting_combos["s2"] = interesting_combos.apply(lambda x: get_strain_i(x, 1), axis = 1)
interesting_combos.head()

## Use Groupby to get relative frequencies of combination results

emergance_events = interesting_combos[["s1", "s2", "treatment_with", "conjugation"]].groupby(
    ["s1", "s2", "treatment_with"]
).sum()
emergance_events = emergance_events.rename(columns={"conjugation":"n"})
emergance_events["total"] = interesting_combos[["s1", "s2", "treatment_with", "conjugation"]].groupby(
    ["s1", "s2", "treatment_with"]
).count()
emergance_events["f"] = round(emergance_events["n"]/emergance_events["total"],2)
emergance_events = emergance_events.reset_index()
emergance_events

### Save for Latex

emergance_events_lt = emergance_events.replace({"A_r^h":"$A^h_r$", "B_r^l":"$B^l_r$", "B_r^h":"$B^h_r$", "A_r^l":"$A^l_r$"}).copy()
emergance_events_lt = emergance_events_lt.rename(columns = {"s1":"sa", "s2":"sb", "treatment_with":"treatment"})
emergance_events_lt.to_latex(os.path.join(experiment.pathes["tables"], "emergence_advanced.tex"))
emergance_events_lt.to_csv(os.path.join(experiment.pathes["tables"], "emergence_advanced.csv"), sep=";")
emergance_events_lt

## Frequency of combinations by strategy
- The result above shows that only turnoverstrains lead to double resistance.
- It is therefore important to know the probabilty that these strains meet by strategy.

freq_encounters = interesting_combos[(interesting_combos.s1 == "A_r^h") & (interesting_combos.s2 == "B_r^h")].copy()
freq_encounters["encounters"] = True
freq_encounters_strat = freq_encounters[["strategy", "encounters"]].groupby("strategy").count()
transfers = len(freq_encounters.transfer_n.unique())
freq_encounters_strat["encounters_per_transfer"] = round(freq_encounters_strat.encounters/transfers, 2)
freq_encounters_strat

## Exclude transfer 1
-  Transfer 1 is special since plate 0 is untreated

freq_encounters_x1 = interesting_combos[(interesting_combos.s1 == "A_r^h") & (interesting_combos.s2 == "B_r^h") & (interesting_combos.transfer_n > 1)].copy()
freq_encounters_x1["encounters_ex1"] = True
freq_encounters_strat_x1 = freq_encounters_x1[["strategy", "encounters_ex1"]].groupby("strategy").count()
transfers_x1 = len(freq_encounters_x1.transfer_n.unique())
freq_encounters_strat_x1["encounters_per_transfer_ex1"] = round(freq_encounters_strat_x1.encounters_ex1/transfers_x1, 2)
freq_encounters_strat_x1

encounters = pd.concat([freq_encounters_strat, freq_encounters_strat_x1], axis=1).fillna(0)
encounters

### For Latex

encounters_lt = encounters.reset_index()
encounters_lt = encounters_lt.rename(columns = {"encounters":"e", "encounters_per_transfer":"fe", "encounters_ex1":"ex", "encounters_per_transfer_ex1":"fex"})
encounters_lt.to_csv(os.path.join(experiment.pathes["tables"], "encounters_advanced.csv"))
encounters_lt

## Encounters between $A_r^h \& B_r^h $
- collect encounters per plate rep unit
    - filter for A_r_h, B_r_h
    - group for strat, t, 
    - f = n/376

groups = ["transfer_n", "rep", "strategy", "n"]

enc_df = interesting_combos.copy()
enc_df["n"] = (enc_df.s1 == "A_r^h") &  (enc_df.s2 == "B_r^h") 
enc_df = enc_df[enc_df.n][groups]
enc_df.head()

enc_sum = enc_df.groupby(["transfer_n", "strategy"]).count().unstack(fill_value=0).stack().reset_index()
enc_sum["f"] = enc_sum["n"]/376
enc_sum["t1"] = enc_sum["transfer_n"] == 1
enc_sum
