==> JupyterScripts/05c_synergy_antha.ipynb <==
from evaluation_scripts.base import get_pathes, load_pickle_obj
jupyter_path, base_path, path, analysis_path, os = get_pathes("20220412_output")
M_path = os.path.join(base_path, "summary", "obj")


M_ab = load_pickle_obj(os.path.join(M_path, "M_t_AB.pkl")) 
M_a  = load_pickle_obj(os.path.join(M_path, "M_t_A.pkl")) 
M_b  = load_pickle_obj(os.path.join(M_path, "M_t_B.pkl")) 

M_a

M_b

M_ab

## Effectivness of Drugs while adding Plasmids (S to Ar or Br)

baseline_A = M_a.loc["U", "S"]
w_A_Br = M_a.loc["U", "B_r"]/baseline_A
print("The effectiveness of A is INCREASED due to plasmid pB by: ", -round(1-w_A_Br, 2)*100, "%")

baseline_B = M_b.loc["U", "S"]
w_B_Ar = M_b.loc["U", "A_r"]/baseline_B
print("The effectiveness of B is reduced due to plasmid pA by: ", round(1-w_B_Ar,2)*100, "%")

baseline_AB = M_ab.loc["U", "S"]
w_AB_Ar = M_ab.loc["U", "A_r"]/baseline_AB
print("The effectiveness of AB is reduced due to plasmid pA by: ", round(1-w_AB_Ar,2)*100, "%")

w_AB_Br = M_ab.loc["U", "B_r"]/baseline_AB
print("The effectiveness of AB is INCREASED due to plasmid pB by: ", -round(1-w_AB_Br,2)*100, "%")

## Effectivness of Drugs on S

e_A_AB = round(1 - baseline_AB/baseline_A, 2)
print("The effectivness of drug AB (", round(baseline_AB, 2), ")is ", e_A_AB*100, " % lower compared to drug A alone(", round(baseline_A, 2), ").")
